playlist changes

+++ src/main/java/com/example/ljudio/repository/PlaylistRepository.java
@@ -3,5 +3,5 @@ package com.example.ljudio.repository;
 import com.example.ljudio.model.Playlist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlaylistRepository extends JpaRepository<Playlist, Integer> {
+public interface PlaylistRepository extends JpaRepository<Playlist, Long> {
 }
